Sanitize the Headers of the AdRenderer response
Some HTTP clients / servers are accepting only a limited set of characters
in the HTTP Headers.

As we are letting our users provbide any value in the displayContext, we
need to sanitize their value before putting it im an HTTP Header.

URL Encoding seems a goog way to do that.