MICS-15536 Rework AudienceFeedConnectorBasePlugin hierarchy
To simplify downstream API, this commit adds 2 new abstract classes to
isolate plugin with or without batching feature.

Thanks to this, it's easy to see in implementation if plugin implements
`AudienceFeedConnectorBasePlugin` or
`BatchedAudienceFeedConnectorBasePlugin<T>`.
When using batch, it also :
 - explicits the type of the line, and of the batch context.
 - require implementation of the onBatchUpdate.

With the vanilla implementation, the implementation cannot respond with
`UserSegmentUpdatePluginBatchDeliveryResponseData`, and cannot override
onBatchUpdate.