MICS-20510 Update Computed Field plugins structure
We need to provide developers with more details when working
on computing fields, such as the type of data they are receiving
(UserActivity, UserProfile or Computed field result updates) but also
the operation that caused the update (insertion or deletion).

To do so, we need to refactor the structure of the computed field
sdk, enforcing the definition of 3 distincts updates method (one per
data type).

Developer experience should be enhanced this way.